requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
attrs==25.3.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
discord.py==2.5.2
distlib==0.3.9
filelock==3.18.0
Flask==3.1.0
frozenlist==1.6.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.4.3
platformdirs==4.3.8
propcache==0.3.1
requests==2.32.3
urllib3==2.4.0
virtualenv==20.31.2
Werkzeug==3.1.3
yarl==1.20.0
python-dotenv==1.0.1
